🚚: Moved resizeImage to /lib
